vlan

+++ layer2/layer2.h
@@ -144,7 +144,7 @@ if(vlan_8021q_hdr->tpid == VLAN_8021Q_PROTO) {
 
 
 static inline unsigned int GET_802_1Q_VLAN_ID(vlan_8021q_hdr_t *vlan_8021q_hdr) {
-    return vlan_8021q_hdr->tci_vid;
+    return (unsigned int)vlan_8021q_hdr->tci_vid;
 }
